weights changed to int, graph import/export, some minor fixes in Dijkstra algorithm

+++ k-shortest-paths-java-version-master/src/main/java/edu/asu/emit/algorithm/graph/VariableGraph.java
@@ -132,7 +132,7 @@ public class VariableGraph extends Graph {
 	 * @param sink
 	 * @return
 	 */
-	public double getEdgeWeight(BaseVertex source, BaseVertex sink)	{
+	public int getEdgeWeight(BaseVertex source, BaseVertex sink)	{
 		int sourceId = source.getId();
 		int sinkId = sink.getId();
