+++ k-shortest-paths-java-version-master/src/main/java/edu/asu/emit/algorithm/graph/abstraction/BaseGraph.java
@@ -45,7 +45,7 @@ public interface BaseGraph {
 	
 	List<BaseVertex> getVertexList();
 	
-	double getEdgeWeight(BaseVertex source, BaseVertex sink);
+	int getEdgeWeight(BaseVertex source, BaseVertex sink);
 	Set<BaseVertex> getAdjacentVertices(BaseVertex vertex);
 	Set<BaseVertex> getPrecedentVertices(BaseVertex vertex);
